requirements.txt
feedparser==5.2.1
python-slugify==1.2.4
